BACKEND/pyproject.toml
[tool.poetry]
name = "fastparking-api"
version = "0.5.28"
description = "GoIT DataScience Team Project - Fast Parking API module"
authors = ["neuron_curva_bobr <info@neuron.curva.bobr.app>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
numpy = "^1.26.4"
opencv-python-headless = {version = "^4.9.0.80", platform = "linux"}
opencv-python = {version = "^4.9.0.80", platform = "win32 || darwin"}
matplotlib = "^3.8.4"
tensorflow = "^2.16.1"
keras = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
